bulk-labeling/pyproject.toml
[build-system]
requires = [
    "setuptools >=65.0",
    "wheel >=0.37",
]
build-backend = "setuptools.build_meta"

[project]
name = "bulk-labeling-solara"
authors = [{name = "Ben Epstein", email = "ben.epstein97@gmail.com"}]
license = {file = "LICENSE"}
dynamic = ["version", "readme"]
classifiers = ["License :: OSI Approved :: Apache License"]
dependencies = [
  "solara",
  "pandas",
  "numpy",
  "plotly<=5.9.0",
  "sentence-transformers",
  "umap-learn",
  "numerize",
]

[project.optional-dependencies]
dev = [
  "jupyter",
  "black",
  "isort",
  "flake8",
]

[project.urls]
Home = "https://www.github.com/ben-epstein/bulk-labeling-solara"

[tool.isort]
profile = "black"


[tool.setuptools.dynamic]
version = {attr = "bulk_labeling.__version__"}

[tool.setuptools]
py-modules = []
